webapp/requirements.txt
flask>=2.3.0
flask-socketio>=5.3.6
torch>=2.0.0
pillow>=9.5.0
numpy>=1.24.0
